Moved the QR class to its own package
