+++ app/src/main/java/com/nervousfish/nervousfish/modules/qr/QRGenerator.java
@@ -1,4 +1,4 @@
-package com.nervousfish.nervousfish;
+package com.nervousfish.nervousfish.modules.qr;
 
 import android.graphics.Bitmap;
 import android.graphics.Color;
@@ -22,10 +22,11 @@ import com.google.zxing.qrcode.QRCodeWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.util.Hashtable;
 
 /**
- * Generates a QR code
+ * Class that can be used to generate QR codes.
  */
 final public class QRGenerator {
 
@@ -34,24 +35,24 @@ final public class QRGenerator {
     private static final int QRCODE_IMAGE_WIDTH = 400;
 
     /**
-     * Unused constructor for utility class
+     * Unused constructor for utility class.
      */
     private QRGenerator() {
-        super();
+        // Prevent instantiations of QRGenerator
     }
 
     /**
      * Encodes the publicKey as a QR code.
      *
-     * @param publicKey The public key that's about to be converted to QR code
-     * @return The bitmap being the QR code
+     * @param publicKey The public key that's about to be converted to QR code.
+     * @return The bitmap being the QR code.
      */
     public static Bitmap encode(final String publicKey) {
         final QRCodeWriter qrWriter = new QRCodeWriter();
-        Bitmap bitmap = null;
+        final Bitmap bitmap = Bitmap.createBitmap(QRCODE_IMAGE_WIDTH, QRCODE_IMAGE_HEIGHT, Bitmap.Config.RGB_565);
+
         try {
             final BitMatrix matrix = qrWriter.encode(publicKey, BarcodeFormat.QR_CODE, QRCODE_IMAGE_WIDTH, QRCODE_IMAGE_HEIGHT);
-            bitmap = Bitmap.createBitmap(QRCODE_IMAGE_WIDTH, QRCODE_IMAGE_HEIGHT, Bitmap.Config.RGB_565);
             for (int x = 0; x < QRCODE_IMAGE_WIDTH; x++) {
                 for (int y = 0; y < QRCODE_IMAGE_HEIGHT; y++) {
                     final int color = (matrix.get(x, y) ? Color.BLACK : Color.WHITE);
@@ -61,33 +62,32 @@ final public class QRGenerator {
         } catch (final WriterException e) {
             LOGGER.error("Failed to encode bitmap", e);
         }
-        return bitmap;
 
+        return bitmap;
     }
 
     /**
-     * Decodes the QRcode to a public key string
+     * Decodes the QRcode to a public key string.
      *
-     * @param QRCode The QR code as a bitmap
+     * @param QRCode The QR code as a bitmap.
      * @return The decoded public key.
+     * @throws IOException If the QR code could not be decoded.
      */
-    public static String decode(final Bitmap QRCode) {
-
+    public static String decode(final Bitmap QRCode) throws IOException {
         final Reader qrReader = new MultiFormatReader();
-
         final int[] intArray = new int[QRCode.getWidth() * QRCode.getHeight()];
-        //copy pixel data from the Bitmap into the 'intArray' array
+
+        // copy pixel data from the Bitmap into the 'intArray' array
         QRCode.getPixels(intArray, 0, QRCode.getWidth(), 0, 0, QRCode.getWidth(), QRCode.getHeight());
 
         final LuminanceSource source = new RGBLuminanceSource(QRCode.getWidth(), QRCode.getHeight(), intArray);
         final BinaryBitmap bitmap = new BinaryBitmap(new HybridBinarizer(source));
-        String publicKey = "";
 
         try {
             final Hashtable<DecodeHintType, Object> decodeHints = new Hashtable<>();
             decodeHints.put(DecodeHintType.PURE_BARCODE, Boolean.TRUE);
             final Result decoded = qrReader.decode(bitmap, decodeHints);
-            publicKey = decoded.getText();
+            return decoded.getText();
         } catch (final NotFoundException e) {
             LOGGER.error("Not Found Exception", e);
         } catch (final ChecksumException e) {
@@ -95,7 +95,8 @@ final public class QRGenerator {
         } catch (final FormatException e) {
             LOGGER.error("Format Exception", e);
         }
-        return publicKey;
+
+        throw new IOException();
     }
 
 }
